integrated backend with frontend and renders links
