+++ Frontend/app/SPL/src/SideBar3.tsx
@@ -1,27 +1,25 @@
 
 import { Link } from "react-router-dom";
-import { SideBar3Data } from "./SideBar3Data";
+//import { SideBar3Data } from "./SideBar3Data";
 import "./styles.css"
 
-function SideBar3() {
+function SideBar3({ responseData }:any) {
   return (
     <div className="SideBarDesign">
       <ul className="nav-menu-items">
-        {SideBar3Data.map((item, index) => 
+        {responseData.map((item:any, index:number) => 
           {return (
-            <div key={index} className={item.className}>
-                <Link to={item.path}>
+            <div key={index} className="nav-text">
+              <text fontSize={12}>{item.category}</text>
+                <Link to={item.link}>
                   <div className="ImageContainer">
                     <img
-                      src={item.image1}
+                      src="./alienware.jpg"
                       alt="Image"
                       width="140px"
                       height="100px"
                     />
                   </div>
-                  {/* {<BarItems>
-                    <span> {item.title} </span>
-          </BarItems>} */}
                 </Link>
             </div>
           )})}
