--- Frontend/app/SPL/src/TextBox.tsx
@@ -1,88 +1,0 @@
-import React, { useState } from "react";
-
-interface Props {
-  displayText: (text: string) => void;
-}
-
-async function makeTask(text:string) {
-  fetch("http://127.0.0.1:5000/start-task",{
-    method:"POST",
-    headers:{
-        "Content-Type":"application/json"
-    },
-    body:JSON.stringify({"text":text})
-}).then((response) => {return response.json()}).then((data) => {
-    return new Promise((resolve) => {
-        const callback = () => { 
-            fetch(`http://127.0.0.1:5000/get-task?job=${data.job}`)
-            .then((res) => {
-                if (res.status === 200){
-                    resolve(res)
-                }
-                else{setTimeout(callback,1000)}  
-            })
-        }
-        callback()
-    })
-}).then((res:any) => {return res.json()}).then((data) => console.log(data)).catch(err => console.log(err.json()));
-}
-
-const TextBox: React.FC<Props> = ({ displayText }) => {
-  const [inputText, setInputText] = useState("");
-
-  const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setInputText(event.target.value);
-  };
-
-  const handleSubmit = () => {
-    makeTask(inputText);
-    setInputText("");
-  };
-
-  return (
-    <div className="mb-3">
-      <label htmlFor="exampleFormControlInput1" className="form-label">
-        Enter Text
-      </label>
-      <input
-        type="text"
-        className="form-control"
-        id="exampleFormControlInput1"
-        placeholder="E.G. Docking Station"
-        value={inputText}
-        onChange={handleChange}
-      />
-      <button className="btn btn-primary" onClick={handleSubmit}>
-        Submit
-      </button>
-    </div>
-  );
-};
-
-export default TextBox;
-
-/*interface props {
-  children: string;
-  onClick: () => void;
-   displayText:  () => void;
-}
-const TextBox = ({ children, onClick, displayText }: props) => {
-  return (
-    <div className="mb-3">
-      <label htmlFor={children} className="form-label">
-        Enter Text
-      </label>
-      <input
-        type="email"
-        className="form-control"
-        id="exampleFormControlInput1"
-        placeholder="E.G. Docking Station"
-      ></input>
-      <button className={"btn btn-primary"} onClick={displayText}>
-        Submit
-      </button>
-    </div>
-  );
-};
-
-export default TextBox; */
