+++ Frontend/app/SPL/src/main.tsx
@@ -1,10 +1,12 @@
 import React from 'react'
+import { BrowserRouter } from 'react-router-dom'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
+  <BrowserRouter><React.StrictMode>
+  <App />
+</React.StrictMode></BrowserRouter>
+  
 )
